use int instead of size_t / unsigned

+++ include/algorithm.h
@@ -9,9 +9,9 @@ class Algorithm
 {
 public:
     Algorithm(std::string name, 
-              unsigned nr_cascade_iterations,
-              size_t (*block_size_function)(unsigned iteration_nr, double estimated_bit_error_rate),
-              unsigned nr_biconf_iterations,
+              int nr_cascade_iterations,
+              int (*block_size_function)(int iteration_nr, double estimated_bit_error_rate),
+              int nr_biconf_iterations,
               bool biconf_error_free_streak,
               bool biconf_correct_complement,
               bool biconf_cascade,
@@ -19,9 +19,9 @@ public:
               bool block_parity_inference);
     static Algorithm *get_by_name(std::string name);
     std::string name;
-    unsigned nr_cascade_iterations;
-    size_t (*block_size_function)(unsigned iteration_nr, double estimated_bit_error_rate);
-    unsigned nr_biconf_iterations;
+    int nr_cascade_iterations;
+    int (*block_size_function)(int iteration_nr, double estimated_bit_error_rate);
+    int nr_biconf_iterations;
     bool biconf_error_free_streak;
     bool biconf_correct_complement;
     bool biconf_cascade;
