+++ include/block.h
@@ -17,15 +17,15 @@ class Block
 {
 public:
     static const int unknown_parity;
-    Block(Iteration& iteration, Block* parent_block, unsigned block_nr, size_t start_bit_nr,
-          size_t end_bit_nr);
+    Block(Iteration& iteration, Block* parent_block, int block_nr, int start_bit_nr,
+          int end_bit_nr);
     ~Block();
     std::string compute_name() const;
-    size_t get_nr_bits() const;
+    int get_nr_bits() const;
     Iteration& get_iteration() const;
-    unsigned get_iteration_nr() const;
-    size_t get_start_bit_nr() const;
-    size_t get_end_bit_nr() const;
+    int get_iteration_nr() const;
+    int get_start_bit_nr() const;
+    int get_end_bit_nr() const;
     int compute_current_parity() const;
     int compute_parity_for_key(const Key& shuffled_key) const;
     void set_correct_parity(int parity);
@@ -38,9 +38,9 @@ public:
     BlockPtr create_right_sub_block();
 private:
     Iteration& iteration;
-    unsigned block_nr;
-    size_t start_bit_nr;
-    size_t end_bit_nr;
+    int block_nr;
+    int start_bit_nr;
+    int end_bit_nr;
     int correct_parity;
     Block* parent_block;
     BlockPtr left_sub_block;
